pyproject.toml
[project]
name = "django-carlog"
version = "2.0.0"
description = "Car trip tracking REST API"
requires-python = ">=3.12"
dependencies = [
    "Django==5.1.5",
    "mysqlclient==2.2.6",
    "django-extensions==3.2.3",
    "django-model-utils==5.0.0",
    "django-environ==0.11.2",
    "djangorestframework==3.15.2",
    "django-filter==24.3",
    "django-allauth==65.3.0",
    "PyJWT==2.10.1",
    "cryptography==44.0.0",
    "gunicorn==23.0.0",
]

[project.optional-dependencies]
dev = [
    "ipython==8.31.0",
    "django-debug-toolbar==4.4.6",
    "pytest==8.3.4",
    "pytest-django==4.9.0",
    "pytest-cov==6.0.0",
    "pytest-playwright==0.6.2",
    "ruff==0.8.6",
    "mypy==1.14.1",
    "django-stubs==5.1.1",
    "pre-commit==4.0.1",
]

[tool.ruff]
target-version = "py312"
line-length = 120
exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "venv",
    "env",
    "migrations",
    "__pycache__",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["trips", "django_carlog"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "migrations/",
    "tests/",
    "e2e/",
]

[tool.django-stubs]
django_settings_module = "django_carlog.settings.base"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_carlog.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["trips"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
